refactor: update bind model - extract placeholder binding to separate classes
