+++ src/NodePathCalculator.php
@@ -5,7 +5,7 @@ use Gt\Dom\Node;
 use Gt\Dom\Facade\NodeClass\DOMElementFacade;
 use Stringable;
 
-class NodePathExtractor implements Stringable {
+class NodePathCalculator implements Stringable {
 	public function __construct(
 		private Node $element
 	) {
