+++ src/PlaceholderBinder.php
@@ -1,19 +1,16 @@
 <?php
 namespace Gt\DomTemplate;
 
-use Gt\Dom\Attr;
 use Gt\Dom\Document;
 use Gt\Dom\Element;
-use Gt\Dom\Text;
 
 class PlaceholderBinder {
-	/** @var array<string, PlaceholderText[]> */
-	private array $placeholderList;
+	private PlaceholderCollection $placeholderCollection;
 
 	public function __construct(
-		private Document $document
+		PlaceholderCollection $placeholderCollection = null
 	) {
-		$this->placeholderList = $this->findPlaceholders($document);
+		$this->placeholderCollection = $placeholderCollection ?? new PlaceholderCollection();
 	}
 
 	public function bind(
@@ -21,56 +18,19 @@ class PlaceholderBinder {
 		mixed $value,
 		Document|Element $context = null
 	):void {
-		if(is_null($context)) {
-			$context = $this->document->documentElement;
+		if($context instanceof Document) {
+			$context = $context->ownerDocument->documentElement;
 		}
 
-		foreach($this->placeholderList[$key] ?? [] as $placeholderText) {
-			if(!$placeholderText->isWithinContext($context)) {
-				continue;
-			}
+		$placeholderList = $this->placeholderCollection->extract($context);
+		foreach($placeholderList as $placeholderTextArray) {
+			foreach($placeholderTextArray as $placeholderText) {
+				if(!$placeholderText->isWithinContext($context)) {
+					continue;
+				}
 
-			$placeholderText->setValue($value);
+				$placeholderText->setKeyValue($key, $value);
+			}
 		}
-	}
-
-	/**
-	 * @return array<string, PlaceholderText[]> An array who's key is the
-	 * bind key and value is an array of matching PlaceholderText objects.
-	 */
-	private function findPlaceholders(Document $document):array {
-		$placeholderList = [];
-
-// The XPath query is split into two, separated by the pipe character (|).
-// The first query: //text()[contains(.,'{{')] finds any Text nodes that contain
-// two opening curly braces.
-// The second query: //@*[contains(.,'{{')] finds any Attr nodes that contain
-// two opening curly braces.
-// NOTE: An Attr node's value is represented by a Text node.
-		$xpathResult = $document->evaluate("//text()[contains(.,'{{')] | //@*[contains(.,'{{')]");
-		foreach($xpathResult as $textOrAttribute) {
-			$text = $textOrAttribute;
-			if($textOrAttribute instanceof Attr) {
-				$text = $textOrAttribute->childNodes[0];
-			}
-
-			/** @var Text $text */
-			$placeholder = $text->splitText(
-				strpos($text->data, "{{")
-			);
-			$placeholder->splitText(
-				strpos($placeholder->data, "}}") + 2
-			);
-
-			$placeholderText = new PlaceholderText($placeholder);
-			$key = $placeholderText->getBindKey();
-			if(!isset($placeholderList[$key])) {
-				$placeholderList[$key] = [];
-			}
-
-			array_push($placeholderList[$key], $placeholderText);
-		}
-
-		return $placeholderList;
 	}
 }
