+++ test/phpunit/PlaceholderBinderTest.php
@@ -7,41 +7,10 @@ use Gt\DomTemplate\Test\TestFactory\DocumentTestFactory;
 use PHPUnit\Framework\TestCase;
 
 class PlaceholderBinderTest extends TestCase {
-	public function testConstructor_noBind():void {
-		$document = DocumentTestFactory::createHTML(DocumentTestFactory::HTML_PLACEHOLDER);
-		$greetingElement = $document->querySelector("#test1 .greeting");
-
-// Before a TemplateCollection is introduced, the placeholders are visible
-// in the source.
-		self::assertSame("Hello, {{name}}!", $greetingElement->textContent);
-		new PlaceholderBinder($document);
-// Now that a TempalteCollection exists, the default text is shown, which
-// in this case is just the bind key, because no other default is supplied.
-		self::assertSame("Hello, name!", $greetingElement->textContent);
-	}
-
-	public function testConstructor_noBind_default():void {
-		$document = DocumentTestFactory::createHTML(DocumentTestFactory::HTML_PLACEHOLDER);
-		$greetingElement = $document->querySelector("#test2 .greeting");
-
-		self::assertSame("Hello, {{name ?? you}}!", $greetingElement->textContent);
-		new PlaceholderBinder($document);
-		self::assertSame("Hello, you!", $greetingElement->textContent);
-	}
-
-	public function testConstructor_noBind_defaultDifferentSyntax():void {
-		$document = DocumentTestFactory::createHTML(DocumentTestFactory::HTML_PLACEHOLDER);
-		$greetingElement = $document->querySelector("#test2a .greeting");
-
-		self::assertSame("Hello, {{name??you}}!", $greetingElement->textContent);
-		new PlaceholderBinder($document);
-		self::assertSame("Hello, you!", $greetingElement->textContent);
-	}
-
 	public function testBind():void {
 		$document = DocumentTestFactory::createHTML(DocumentTestFactory::HTML_PLACEHOLDER);
 		$greetingElement = $document->querySelector("#test2 .greeting");
-		$sut = new PlaceholderBinder($document);
+		$sut = new PlaceholderBinder();
 // We can now bind text to the placeholder, and the text will
 // magically be replaced.
 		$sut->bind("name", "Cody", $greetingElement);
@@ -52,7 +21,7 @@ class PlaceholderBinderTest extends TestCase {
 	public function testBind_contextDoesNotLeak():void {
 		$document = DocumentTestFactory::createHTML(DocumentTestFactory::HTML_PLACEHOLDER);
 		$greetingElement = $document->querySelector("#test2 .greeting");
-		$sut = new PlaceholderBinder($document);
+		$sut = new PlaceholderBinder();
 		$sut->bind("name", "Cody", $greetingElement);
 		self::assertStringContainsString("Cody", $document->querySelector("#test2 .greeting")->textContent);
 		self::assertStringNotContainsString("Cody", $document->querySelector("#test1 .greeting")->textContent);
@@ -61,8 +30,8 @@ class PlaceholderBinderTest extends TestCase {
 
 	public function testBind_noContextBindsAll():void {
 		$document = DocumentTestFactory::createHTML(DocumentTestFactory::HTML_PLACEHOLDER);
-		$sut = new PlaceholderBinder($document);
-		$sut->bind("name", "Cody");
+		$sut = new PlaceholderBinder();
+		$sut->bind("name", "Cody", $document);
 		self::assertStringContainsString("Cody", $document->querySelector("#test1 .greeting")->textContent);
 		self::assertStringContainsString("Cody", $document->querySelector("#test2 .greeting")->textContent);
 		self::assertStringContainsString("Cody", $document->querySelector("#test2a .greeting")->textContent);
@@ -70,27 +39,11 @@ class PlaceholderBinderTest extends TestCase {
 
 	public function testBind_attribute():void {
 		$document = DocumentTestFactory::createHTML(DocumentTestFactory::HTML_PLACEHOLDER);
-		$sut = new PlaceholderBinder($document);
+		$sut = new PlaceholderBinder();
 		$testElement = $document->getElementById("test3");
 		/** @var HTMLAnchorElement $link */
 		$link = $testElement->querySelector("a");
 		$sut->bind("repoName", "domtemplate", $testElement);
 		self::assertSame("https://www.php.gt/domtemplate", $link->href);
-	}
-
-	public function testBind_attributeDefault():void {
-		$document = DocumentTestFactory::createHTML(DocumentTestFactory::HTML_PLACEHOLDER);
-		$testElement = $document->getElementById("test4");
-		/** @var HTMLAnchorElement $link */
-		$link = $testElement->querySelector("a");
-		self::assertSame(
-			"https://www.php.gt/{{repoName ?? domtemplate}}",
-			$link->href
-		);
-		new PlaceholderBinder($document);
-		self::assertSame(
-			"https://www.php.gt/domtemplate",
-			$link->href
-		);
 	}
 }
